refactor(tests): consolidate orchestrator tests from 17 -> 12 subsystem suites
Merge nine fragmented orchestrator test files into three feature-level
suites without dropping any behavioral coverage:

- orchestratorAdapters.test.ts <- baseOrchestratorAdapter +
  providerOrchestratorAdapter + permissionMapping + modelConfigResolver
  (13 it cases)
- mission.test.ts <- missionLifecycle + missionBudgetService +
  missionStateDoc (35 it cases)
- orchestratorPlanning.test.ts <- orchestratorContext +
  delegationContracts (14 it cases)

Kept separate (genuinely different concerns / file size):
orchestratorService, aiOrchestratorService, executionPolicy,
promptInspector, adaptiveRuntime, runtimeEventRouter,
workerDeliveryService, coordinatorAgent, coordinatorTools.

All 463 tests in the consolidated orchestrator suite pass.
